requirements.txt
PyYAML>=6.0.2
fastapi>=0.116.0
uvicorn>=0.35.0
ollama>=0.5.1
vllm>=0.16.0
vllm-metal @ git+https://github.com/vllm-project/vllm-metal.git
